Fixed problems with login in web version and changes in Conversacion and Anuncio
Modified implementation of the private access in Conversacion and
Anuncio. Solved problems with login in web part. When a request is made
in postman with login user required, returns a 403 error.

+++ wallacomic/src/main/java/com/wallacomic/api/AnuncioRestController.java
@@ -40,7 +40,7 @@ public class AnuncioRestController {
 		}
 		
 	}
-	//OJO!!! hay que poner las urls como peticiones de recursos, no con nombres de métodos
+
 	@RequestMapping(value = "/", method = RequestMethod.POST)
 	@ResponseStatus(HttpStatus.CREATED)
 	public Anuncio createAdvertisement(@RequestBody Anuncio ad) {
@@ -53,8 +53,12 @@ public class AnuncioRestController {
 	@RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
 	public ResponseEntity<Anuncio> deleteAdvertisement(@PathVariable long id) {
 		
-		anuncioService.delete(id);
-		return new ResponseEntity<>(null, HttpStatus.OK);
+		if(anuncioService.existAd(id)){
+			anuncioService.delete(id);
+			return new ResponseEntity <>(null, HttpStatus.OK);
+		}else{
+			return new ResponseEntity <>(HttpStatus.NOT_FOUND);
+		}
 		
 	}
