+++ wallacomic/src/main/java/com/wallacomic/service/AnuncioService.java
@@ -3,16 +3,21 @@ package com.wallacomic.service;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.stereotype.Service;
 
 import com.wallacomic.domain.Anuncio;
 import com.wallacomic.domain.Comic;
 import com.wallacomic.domain.Usuario;
+import com.wallacomic.domain.UsuarioComponent;
 import com.wallacomic.repository.AnuncioRepository;
 
 
 @Service
 public class AnuncioService {
+	
+	@Autowired
+	UsuarioComponent usuarioComponent;
 
 	@Autowired
 	private AnuncioRepository anuncioRepository;
@@ -30,7 +35,13 @@ public class AnuncioService {
 	}
 	
 	public void save(Anuncio ad){
-		anuncioRepository.save(ad);
+		
+		if(usuarioComponent.isLoggedUser() && usuarioComponent.hasAdminPermissions()){
+			anuncioRepository.save(ad);
+		}else{
+			throw new BadCredentialsException("Error de creacion");
+		}
+		
 	}
 	
 	public Anuncio findById(long id){
@@ -38,11 +49,18 @@ public class AnuncioService {
 	}
 	
 	public void delete(long id){
-		anuncioRepository.delete(id);
+		
+		if(usuarioComponent.isLoggedUser() && usuarioComponent.hasAdminPermissions()){
+			anuncioRepository.delete(id);
+		}else{
+			throw new BadCredentialsException("Error de creacion");
+		}
+		
 	}
 	
 	public boolean existAd(long id){
 		return anuncioRepository.exists(id);
 	}
+	
 	
 }
